Switched to Room DB
Successfully got vehicle to enter DB

Can enter both at the same time but need to fix mileage input via popup

+++ app/src/main/java/com/staxxproducts/vehicleservicetracker/data/VehicleStudentCrossRef.kt
@@ -4,6 +4,6 @@ import androidx.room.Entity
 
 @Entity(primaryKeys = ["vehicleId","serviceId"])
 data class VehicleStudentCrossRef(
-    val vehicleId: Int,
-    val serviceId: Int
+    val vehicleId: Long,
+    val serviceId: Long
 )
